API

LibPQ API

LibPQ.Connection — Type.

type Connection

A connection to a PostgreSQL database.

Fields:

conn

A pointer to a libpq PGconn object (C_NULL if closed)

closed

True if the connection is closed and the PGconn object has been cleaned up

source

LibPQ.Connection — Method.

Connection(str::AbstractString; throw_error=true) -> Connection

Create a Connection from a connection string as specified in the PostgreSQL documentation (33.1.1. Connection Strings). If throw_error is true, an error will be thrown if the resulting connection's status is CONNECTION_BAD and the PGconn object will be cleaned up. Otherwise, a warning will be shown and the user should call close or reset! on the returned Connection.

source

LibPQ.clear — Method.

clear(jl_result::Result)

Clean up the memory used by the PGresult object. The Result will no longer be usable.

source

LibPQ.execute — Method.

execute(jl_conn::Connection, query::AbstractString; throw_error=false) -> Result

Run a query on the PostgreSQL database and return a Result. If throw_error is true, throw an error and clear the result if the query results in a fatal error or unreadable response.

source

LibPQ.status — Method.

status(jl_conn::Connection) -> libpq_c.ConnStatusType

Return the status of the PostgreSQL database connection according to libpq. Only CONNECTION_OK and CONNECTION_BAD are valid for blocking connections, and only blocking connections are supported right now.

See also: error_message

source

LibPQ.status — Method.

status(jl_result::Result) -> libpq_c.ExecStatusType

Return the status of a result's corresponding database query according to libpq. Only CONNECTION_OK and CONNECTION_BAD are valid for blocking connections, and only blocking connections are supported right now.

See also: error_message

source

LibPQ.ConnectionOption — Type.

struct ConnectionOption

A Julia representation of a PostgreSQL connection option (PQconninfoOption).

Fields:

keyword

The name of the option

envvar

The name of the fallback environment variable for this option

compiled

The PostgreSQL compiled-in default for this option

val

The value of the option if set

label

The label of the option for display

disptype

Indicator for how to display the option (see ConninfoDisplay)

dispsize

The size of field to provide for entry of the option value (not used here)

source

LibPQ.ConnectionOption — Method.

ConnectionOption(pq_opt::libpq_c.PQconninfoOption) -> ConnectionOption

Construct a ConnectionOption from a libpg_c.PQconninfoOption.

source

LibPQ.ConninfoDisplay — Type.

Indicator for how to display a PostgreSQL connection option (PQconninfoOption).

Possible values are:

Normal (libpq: ""): display as is

Password (libpq: "*"): hide the value of this field

Debug (libpq: "D"): don't show by default

source

LibPQ.Result — Type.

type Result <: DataStreams.Data.Source

A result from a PostgreSQL database query

Fields:

result

A pointer to a libpq PGresult object (C_NULL if cleared)

cleared

True if the PGresult object has been cleaned up

source

LibPQ.Result — Method.

Result(result::Ptr{libpq_c.PGresult}) -> Result

Construct a Result from a libpg_c.PGresult

source

Base.close — Method.

close(jl_conn::Connection)

Close the PostgreSQL database connection and free the memory used by the PGconn object. This function calls PQfinish, but only if jl_conn.closed is false, to avoid a double-free.

source

Base.parse — Method.

parse(::Type{ConninfoDisplay}, str::AbstractString) -> ConninfoDisplay

Parse a ConninfoDisplay from a string. See ConninfoDisplay.

source

Base.show — Method.

show(io::IO, jl_conn::Connection)

Display a Connection by showing the connection status and each connection option.

source

Base.show — Method.

show(io::IO, jl_result::Result)

Show a PostgreSQL result and whether it has been cleared.

source

LibPQ.column_name — Method.

column_name(jl_result::Result, column_number::Integer) -> String

Return the name of the column at index column_number (1-based).

source

LibPQ.column_names — Method.

column_names(jl_result::Result, column_number::Integer) -> Vector{String}

Return the names of all the columns in the query result.

source

LibPQ.column_number — Method.

column_number(jl_result::Result, column_name::AbstractString) -> Int

Return the index (1-based) of the column named column_name.

source

LibPQ.conninfo — Method.

conninfo(jl_conn::Connection) -> Vector{ConnectionOption}

Get all connection options for a connection.

source

LibPQ.error_message — Method.

error_message(jl_conn::Connection) -> String

Return the error message most recently generated by an operation on the connection. Includes a trailing newline.

source

LibPQ.error_message — Method.

error_message(jl_result::Result) -> String

Return the error message associated with the result, or an empty string if there was no error. Includes a trailing newline.

source

LibPQ.num_columns — Method.

num_columns(jl_result::Result) -> Int

Return the number of columns in the query result. This will be 0 if the query would never return data.

source

LibPQ.num_rows — Method.

num_rows(jl_result::Result) -> Int

Return the number of rows in the query result. This will be 0 if the query would never return data.

source

LibPQ.reset! — Method.

reset!(jl_conn::Connection; throw_error=true)

Reset the communication to the PostgreSQL server. The PGconn object will be recreated using identical connection parameters. The throw_error parameter functions as in Connection.

Note

This function can be called on a connection with status CONNECTION_BAD, for example, but cannot be called on a connection that has been closed.

source

LibPQ.unsafe_nullable_string — Method.

unsafe_nullable_string(ptr::Cstring) -> Nullable{String}

Convert a Cstring to a Nullable{String}, returning Nullable{String}() if the pointer is C_NULL.

source