API

LibPQ API

LibPQ.Connection — Type.

type Connection

A connection to a PostgreSQL database.

Fields:

conn

A pointer to a libpq PGconn object (C_NULL if closed)

closed

True if the connection is closed and the PGconn object has been cleaned up

encoding

libpq client encoding (string encoding of returned data)

uid_counter

Integer counter for generating connection-level unique identifiers

source

LibPQ.Connection — Method.

Connection(str::AbstractString; throw_error=true) -> Connection

Create a Connection from a connection string as specified in the PostgreSQL documentation (33.1.1. Connection Strings).

See handle_new_connection for information on the throw_error argument.

source

LibPQ.Result — Type.

type Result <: DataStreams.Data.Source

A result from a PostgreSQL database query

Fields:

result

A pointer to a libpq PGresult object (C_NULL if cleared)

cleared

True if the PGresult object has been cleaned up

source

LibPQ.Result — Method.

Result(result::Ptr{libpq_c.PGresult}) -> Result

Construct a Result from a libpg_c.PGresult

source

LibPQ.Statement — Type.

struct Statement

A PostgreSQL prepared statement

Fields:

jl_conn

A Connection for which this statement is valid. It may become invalid if the connection is reset.

name

An autogenerated neame for the prepared statement (using unique_id

description

A Result containing a description of the prepared statement

num_params

The number of parameters accepted by this statement according to description

source

LibPQ.column_name — Method.

column_name(jl_result::Result, column_number::Integer) -> String

Return the name of the column at index column_number (1-based).

source

LibPQ.column_name — Method.

column_name(jl_result::Result, column_number::Integer) -> String

Return the name of the column at index column_number (1-based) that would be returned by executing the prepared statement.

source

LibPQ.column_names — Method.

column_names(jl_result::Result, column_number::Integer) -> Vector{String}

Return the names of all the columns in the query result.

source

LibPQ.column_names — Method.

column_names(jl_result::Result, column_number::Integer) -> Vector{String}

Return the names of all the columns in the query result that would be returned by executing the prepared statement.

source

LibPQ.column_number — Method.

column_number(jl_result::Result, column_name::AbstractString) -> Int

Return the index (1-based) of the column named column_name.

source

LibPQ.column_number — Method.

column_number(jl_result::Result, column_name::AbstractString) -> Int

Return the index (1-based) of the column named column_name that would be returned by executing the prepared statement.

source

LibPQ.encoding — Method.

encoding(jl_conn::Connection) -> String

Return the client encoding name for the current connection (see Table 23.1. PostgreSQL Character Sets for possible values).

Currently all Julia connections are set to use UTF8 as this makes conversion to and from String straighforward.

See also: set_encoding!, reset_encoding!

source

LibPQ.execute — Method.

execute(jl_conn::Connection, query::AbstractString, parameters::Vector{<:Parameter}; throw_error=true) -> Result

Run a query on the PostgreSQL database and return a Result. If throw_error is true, throw an error and clear the result if the query results in a fatal error or unreadable response.

source

LibPQ.execute — Method.

execute(jl_conn::Connection, query::AbstractString; throw_error=true) -> Result

Run a query on the PostgreSQL database and return a Result. If throw_error is true, throw an error and clear the result if the query results in a fatal error or unreadable response.

source

LibPQ.execute — Method.

execute(stmt::Statement, parameters::Vector{<:Parameter}; throw_error=true) -> Result

Execute a prepared statement on the PostgreSQL database and return a Result. If throw_error is true, throw an error and clear the result if the query results in a fatal error or unreadable response.

source

LibPQ.num_columns — Method.

num_columns(jl_result::Result) -> Int

Return the number of columns in the query result. This will be 0 if the query would never return data.

source

LibPQ.num_columns — Method.

num_columns(stmt::Statement) -> Int

Return the number of columns that would be returned by executing the prepared statement.

source

LibPQ.num_params — Method.

num_params(jl_result::Result) -> Int

Return the number of parameters in a prepared statement. If this result did not come from the description of a prepared statement, return 0.

source

LibPQ.num_params — Method.

num_params(stmt::Statement) -> Int

Return the number of parameters in the prepared statement.

source

LibPQ.num_rows — Method.

num_rows(jl_result::Result) -> Int

Return the number of rows in the query result. This will be 0 if the query would never return data.

source

LibPQ.reset! — Method.

reset!(jl_conn::Connection; throw_error=true)

Reset the communication to the PostgreSQL server. The PGconn object will be recreated using identical connection parameters.

See handle_new_connection for information on the throw_error argument.

Note

This function can be called on a connection with status CONNECTION_BAD, for example, but cannot be called on a connection that has been closed.

source

LibPQ.reset_encoding! — Method.

reset_encoding!(jl_conn::Connection, encoding::String)

Reset the client encoding for the current connection to jl_conn.encoding.

See also: encoding, set_encoding!

source

LibPQ.server_version — Method.

server_version(jl_conn::Connection) -> VersionNumber

Get the PostgreSQL version of the server.

See 33.2. Connection Status Functions for information on the integer returned by PQserverVersion that is parsed by this function.

See @pqv_str for information on how this packages represents PostgreSQL version numbers.

source

LibPQ.set_encoding! — Method.

set_encoding!(jl_conn::Connection, encoding::String)

Set the client encoding for the current connection (see Table 23.1. PostgreSQL Character Sets for possible values).

Currently all Julia connections are set to use UTF8 as this makes conversion to and from String straighforward. Other encodings are not explicitly handled by this package and will probably be very buggy.

See also: encoding, reset_encoding!

source

LibPQ.status — Method.

status(jl_conn::Connection) -> libpq_c.ConnStatusType

Return the status of the PostgreSQL database connection according to libpq. Only CONNECTION_OK and CONNECTION_BAD are valid for blocking connections, and only blocking connections are supported right now.

See also: error_message

source

LibPQ.status — Method.

status(jl_result::Result) -> libpq_c.ExecStatusType

Return the status of a result's corresponding database query according to libpq. Only CONNECTION_OK and CONNECTION_BAD are valid for blocking connections, and only blocking connections are supported right now.

See also: error_message

source

LibPQ.@pqv_str — Macro.

@pqv_str -> VersionNumber

Parse a PostgreSQL version.

Note

As of version 10.0, PostgreSQL moved from a three-part version number (where the first two parts represent the major version and the third represents the minor version) to a two-part major-minor version number. In LibPQ.jl, we represent this using the first two VersionNumber components as the major version and the third as the minor version.

Examples

julia> pqv"10.1" == v"10.0.1"
true

julia> pqv"9.2.5" == v"9.2.5"
true

source

LibPQ.Statement — Method.

Statement(sch::Data.Schema, ::Type{Data.Row}, append, connection::Connection, query::AbstractString) -> Statement

Construct a Statement for use in streaming with DataStreams. This function is called by Data.stream!(source, Statement, connection, query).

source

LibPQ.fetch! — Method.

fetch!(sink::Union{T, Type{T}}, result::Result, args...; kwargs...) where {T} -> T

Stream data to sink or a new structure of type T using Data.stream!. Any trailing args or kwargs are passed to Data.stream!. result is cleared upon completion.

source

LibPQ.ConnectionOption — Type.

struct ConnectionOption

A Julia representation of a PostgreSQL connection option (PQconninfoOption).

Fields:

keyword

The name of the option

envvar

The name of the fallback environment variable for this option

compiled

The PostgreSQL compiled-in default for this option

val

The value of the option if set

label

The label of the option for display

disptype

Indicator for how to display the option (see ConninfoDisplay)

dispsize

The size of field to provide for entry of the option value (not used here)

source

LibPQ.ConnectionOption — Method.

ConnectionOption(pq_opt::libpq_c.PQconninfoOption) -> ConnectionOption

Construct a ConnectionOption from a libpg_c.PQconninfoOption.

source

LibPQ.ConninfoDisplay — Type.

Indicator for how to display a PostgreSQL connection option (PQconninfoOption).

Possible values are:

Normal (libpq: ""): display as is

Password (libpq: "*"): hide the value of this field

Debug (libpq: "D"): don't show by default

source

Base.Distributed.clear! — Method.

clear!(jl_result::Result)

Clean up the memory used by the PGresult object. The Result will no longer be usable.

source

Base.close — Method.

close(jl_conn::Connection)

Close the PostgreSQL database connection and free the memory used by the PGconn object. This function calls PQfinish, but only if jl_conn.closed is false, to avoid a double-free.

source

Base.isopen — Method.

isopen(jl_conn::Connection) -> Bool

Check whether a connection is open

source

Base.parse — Method.

parse(::Type{ConninfoDisplay}, str::AbstractString) -> ConninfoDisplay

Parse a ConninfoDisplay from a string. See ConninfoDisplay.

source

Base.show — Method.

show(io::IO, jl_conn::Connection)

Display a Connection by showing the connection status and each connection option.

source

Base.show — Method.

show(io::IO, jl_result::Result)

Show a PostgreSQL result and whether it has been cleared.

source

LibPQ.conninfo — Method.

conninfo(jl_conn::Connection) -> Vector{ConnectionOption}

Get all connection options for a connection.

source

LibPQ.error_message — Method.

error_message(jl_conn::Connection) -> String

Return the error message most recently generated by an operation on the connection. Includes a trailing newline.

source

LibPQ.error_message — Method.

error_message(jl_result::Result) -> String

Return the error message associated with the result, or an empty string if there was no error. Includes a trailing newline.

source

LibPQ.handle_new_connection — Method.

handle_new_connection(jl_conn::Connection; throw_error=true) -> Connection

Check status and handle errors for newly-created connections. Also set the client encoding (23.3. Character Set Support) to jl_conn.encoding.

If throw_error is true, an error will be thrown if the connection's status is CONNECTION_BAD and the PGconn object will be cleaned up. Otherwise, a warning will be shown and the user should call close or reset! on the returned Connection.

source

LibPQ.handle_result — Method.

handle_result(jl_result::Result; throw_error::Bool=true) -> Result

Check status and handle errors for newly-created result objects.

If throw_error is true, throw an error and clear the result if the query results in a fatal error or unreadable response. Otherwise a warning is shown.

Also print an info message about the result.

source

LibPQ.prepare — Method.

prepare(jl_conn::Connection, query::AbstractString) -> Statement

Create a prepared statement on the PostgreSQL server using libpq. The statement is given an generated unique name using unique_id.

Note

Currently the statement is not explicitly deallocated, but it is deallocated at the end of session per the PostgreSQL documentation on DEALLOCATE.

source

LibPQ.transaction_status — Method.

transaction_status(jl_conn::Connection) -> libpq_c.PGTransactionStatusType

Return the PostgreSQL database server's current in-transaction status for the connection. See for information on the meaning of the possible return values.

source

LibPQ.unique_id — Function.

unique_id(jl_conn::Connection, prefix::AbstractString="") -> String

Return a valid PostgreSQL identifier that is unique for the current connection. This is mostly used to create names for prepared statements.

source

LibPQ.unsafe_string_or_null — Method.

unsafe_string_or_null(ptr::Cstring) -> Union{String, Missing}

Convert a Cstring to a Union{String, Missing}, returning missing if the pointer is C_NULL.

source